requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
pyyaml>=6.0
docker>=6.0.0
pathlib2>=2.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
tabulate>=0.8.0